obsync: Parallelize Obsync
Using the multiprocessing library for parallelism, the tempfile library
for temporary output targets allowing output collation, and the logging
library to more precisely control the boto logger, obsync can now sync
objects in parallel, while producing debug output as readable as if they
had been done serially.

This patch does not yet handle tunable toleration of transient failures.

Great length is taken to be UTF-8 compliant in this patch, so logging
and tempfiles are made to use UTF-8 where possible, as is the sys module
by way of reloading it to use sys.setdefaultencoding()

The multiprocessing.pool IMapIterator is wrapped to have a timeout so
that KeyboardInterrupts work properly.  Also the pid is included in
tempfile names so that tempfiles can be easily identified and cleaned up
if left behind.