obsync: Tuneable failure toleration
Add a command line argument --tolerate-failures which defaults to 0 but
may be set to a positive integer to allow a the specified number of
Temporary errors to occur without aborting the migration.  The migration
is still considered failed if there have been any errors, but this
allows obsync to be more robust in the face of transient error-causing
conditions.

If any failures were tolerated, list them whether or not the migration
reached completion.  The last tolerated failure is reraised if the
migration reached completion.